Add more test cases (bug fix for 8 bit dac)

+++ src/dac.c
@@ -2,7 +2,7 @@
 
 static uint16_t align_dac_input_data(enum DacDataAlignment dac_align, uint16_t data)
 {
-	uint16_t aligned_data = data;
+	uint16_t aligned_data = (uint8_t) data;
 
 	if (dac_align == TwelveBitRight) {
 		aligned_data = data & 0xFFF;
